user auth added

+++ ranthub/ranthubapp/views.py
@@ -3,7 +3,42 @@ from django.db.models import Q
 from django.http import HttpResponse
 from .models import Room, Topic
 from .forms import RoomForm
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+
 # Create your views here.
+def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+       
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        
+        else:
+            messages.error(request, 'Username or Password is incorrect')
+
+    context = {}
+    return render(request, 'ranthubapp/login.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -21,6 +56,7 @@ def room(request ,pk):
     context = {'room':room}
     return render(request, 'ranthubapp/room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -33,10 +69,14 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'ranthubapp/room_form.html',context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -45,7 +85,11 @@ def updateRoom(request, pk):
     context = {'form':form}
     return render(request, 'ranthubapp/room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+    
     room = Room.objects.get(id=pk)
     if request.method == 'POST':
         room.delete()
